feat: 쿠폰 페이지 UI 수정 완료 (피그마 버전에 맞게) #13
